// types/enum.ts
export enum InputTypes {
  Input = "input",
  Textarea = "textarea",
  DatePicker = "datepicker",
  Select = "select",
  Checkbox = "checkbox",
  Phone_input = "phone_input",
  Radio = "radio",
  UploadFile = "uploadfile",
}

export enum AdminCardType {
  appointment = "scheduled",
  pending = "pending",
  cancelled = "cancelled",
}